# apps/main-platform/app/lib/contracts/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CorpSec Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/redteam/connections": {
      "post": {
        "tags": [
          "redteam"
        ],
        "summary": "Create Connection",
        "operationId": "create_connection_redteam_connections_post",
        "parameters": [
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedTeamConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedTeamConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "List Connections",
        "operationId": "list_connections_redteam_connections_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedTeamConnection"
                  },
                  "title": "Response List Connections Redteam Connections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/redteam/runs": {
      "post": {
        "tags": [
          "redteam"
        ],
        "summary": "Create Run",
        "operationId": "create_run_redteam_runs_post",
        "parameters": [
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedTeamRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedTeamRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_redteam_runs_get",
        "parameters": [
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedTeamRun"
                  },
                  "title": "Response List Runs Redteam Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/redteam/runs/{run_id}": {
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "Get Run",
        "operationId": "get_run_redteam_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedTeamRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/redteam/runs/{run_id}/report": {
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "Get Run Report",
        "operationId": "get_run_report_redteam_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/redteam/runs/{run_id}/events": {
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "Stream Run Events",
        "operationId": "stream_run_events_redteam_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/redteam/runs/{run_id}/evidence": {
      "get": {
        "tags": [
          "redteam"
        ],
        "summary": "Get Run Evidence",
        "operationId": "get_run_evidence_redteam_runs__run_id__evidence_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-redteam-owner",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Owner"
            }
          },
          {
            "name": "x-redteam-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Redteam-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRedTeamConnectionRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Agent Id"
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Endpoint"
          },
          "auth_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Reference"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "endpoint"
        ],
        "title": "CreateRedTeamConnectionRequest"
      },
      "CreateRedTeamRunRequest": {
        "properties": {
          "connection_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Connection Id"
          },
          "config": {
            "$ref": "#/components/schemas/RedTeamRunConfig"
          }
        },
        "type": "object",
        "required": [
          "connection_id"
        ],
        "title": "CreateRedTeamRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RedTeamConnection": {
        "properties": {
          "connection_id": {
            "type": "string",
            "title": "Connection Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "adapter_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Adapter Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "connection_id",
          "agent_id",
          "endpoint",
          "created_at"
        ],
        "title": "RedTeamConnection"
      },
      "RedTeamRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "connection_id": {
            "type": "string",
            "title": "Connection Id"
          },
          "random_seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Random Seed"
          },
          "adapter_metadata_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Adapter Metadata Snapshot"
          },
          "config": {
            "$ref": "#/components/schemas/RedTeamRunConfig"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "current_round": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current Round",
            "default": 0
          },
          "last_event_seq": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Last Event Seq",
            "default": 0
          },
          "report_available": {
            "type": "boolean",
            "title": "Report Available",
            "default": false
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "selected_seed_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Seed Ids"
          },
          "selected_seed_snapshot": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Selected Seed Snapshot"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "agent_id",
          "connection_id",
          "random_seed",
          "config",
          "status",
          "created_at"
        ],
        "title": "RedTeamRun"
      },
      "RedTeamRunConfig": {
        "properties": {
          "rounds": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Rounds",
            "default": 2
          },
          "seed_count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Seed Count",
            "default": 4
          },
          "variants_per_seed": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Variants Per Seed",
            "default": 2
          }
        },
        "type": "object",
        "title": "RedTeamRunConfig"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}